rename normalizeThemeUI to normalizeThemeUIColors
