+++ themes/gatsby-theme-specimens/index.js
@@ -4,9 +4,20 @@ import ColorRow from "./src/components/color-row"
 import ColorSwatch from "./src/components/color-swatch"
 import Download from "./src/components/download"
 import filterStyles from "./src/utils/filter-styles"
-import normalizeThemeUI from "./src/utils/normalize-theme-ui"
+import normalizeThemeUIColors from "./src/utils/normalize-theme-ui-colors"
 import Palette from "./src/components/palette"
 import Typography from "./src/components/typography"
 import Video from "./src/components/video"
 
-export { Audio, Alert, ColorRow, ColorSwatch, Download, filterStyles, normalizeThemeUI, Palette, Typography, Video }
+export {
+  Audio,
+  Alert,
+  ColorRow,
+  ColorSwatch,
+  Download,
+  filterStyles,
+  normalizeThemeUIColors,
+  Palette,
+  Typography,
+  Video,
+}
